requirements.txt
# Django and Related Packages
Django==3.2.4
django-widget-tweaks==1.4.8
django-dbbackup==3.14.0
django-otp==1.1.3
django-filter==2.4.0
django-import-export==2.7.0

# External Packages
openpyxl==3.0.12
qrcode[pil]==7.3
celery==5.1.2

# Miscellaneous Packages
chardet==4.0.0
django-multiupload==0.10.0
